requirement.txt
--index-url https://pypi.org/simple/
--extra-index-url https://test.pypi.org/simple/
Django==4.2.11
psycopg2-binary==2.9.9
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
requests==2.22.0
pymemcache==4.0.0
authorization-package-fkb==0.0.4
